+++ Source/Defcon/GameObjects/Enemies/spacehum.cpp
@@ -94,7 +94,7 @@ void Defcon::CSpacehum::Explode(CGameObjectCollection& debris)
 	m_fLifespan = 0.0f;
 	this->OnAboutToDie();
 
-	for(size_t i = 0; i < 10; i++)
+	for(int32 i = 0; i < 10; i++)
 	{
 		CFlak* pFlak = new CFlak;
 		pFlak->m_eColorbaseYoung = cby;
