+++ Source/Defcon/GameObjects/Enemies/turret.cpp
@@ -194,7 +194,7 @@ void Defcon::CTurret::Explode(CGameObjectCollection& debris)
 	m_fLifespan = 0.0f;
 	this->OnAboutToDie();
 
-	for(size_t i = 0; i < 100; i++)
+	for(int32 i = 0; i < 100; i++)
 	{
 		CFlak* pFlak = new CFlak;
